% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{as.data.frame.parray}
\alias{as.data.frame.parray}
\title{Convert a parray to a data frame}
\usage{
\method{as.data.frame}{parray}(x, row.names = NULL, optional, responseName = "(Weight)", ...)
}
\arguments{
\item{x}{A parray}

\item{row.names}{The names of the rows}

\item{optional}{ignored}

\item{responseName}{The name of the variable}

\item{...}{currently ignored}
}
\value{
A data frame
}
\description{
Convert a parray to a data frame.
}
\details{
Convert a parray to a data frame.
}
