% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{condition}
\alias{condition}
\title{Conditioning probability arrays}
\usage{
condition(P, by = c(), fill = 0)
}
\arguments{
\item{P}{A named array of weights or probabilities}

\item{by}{A vector of variable names to condition on}

\item{fill}{A vector or named array of replacement values for
division by zero.}
}
\value{
An array of conditional probabilities
}
\description{
Condition an array of weights or probabilities on a given
set of variables.
}
\details{
This function conditions an array of weights or probabilities on a
given set of variables.  The array is permuted so that the
conditioning variables are the first dimensions of the array, and
then attempts to normalize so that for each combination of the
conditioning variables the corresponding array slice sums to one.
Where this normalization results in division by zero, the
corresponding values are replaced with the element of "fill",
which may be a vector or named array.
}
