% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{marginalize}
\alias{marginalize}
\title{Marginalizing probability arrays}
\usage{
marginalize(P, to)
}
\arguments{
\item{P}{A named array of weights or probabilities}

\item{to}{A vector of variable names to marginalize to}
}
\value{
An array of marginal probabilities
}
\description{
Marginalize an array of weights or probabilities to a given
set of variables.
}
\details{
Marginalize an array of weights or probabilities, summing over
any dependent variables not given in \code{to}.  Any conditioning
variables are retained.  If \code{prob} is \code{TRUE}, the result is
normalized so that for each combination of the conditioning
variables, the sum over the variables in \code{to} is one.
}
