% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray-package.R
\docType{package}
\name{parray-package}
\alias{parray}
\alias{parray-package}
\title{parray: Arrays of conditional proportions}
\description{
Manipulate arrays of conditional probabilities.
}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
