% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{parse_cformula}
\alias{parse_cformula}
\title{Parse a conditional formula}
\usage{
parse_cformula(formula)
}
\arguments{
\item{formula}{A formula}
}
\value{
A list with elements rvar, dvars, and cvars
}
\description{
Parse a conditional formula
}
\details{
Parse a formula of the form \code{y ~ d_1+d_2+...+d_n | c_1+c_2+...+c_m}
to extract the response variable \code{y}, dependent variables \code{d_i},
and conditioning variables \code{c_i} as character vectors.
}
