% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{product}
\alias{product}
\title{Product of two arrays}
\usage{
product(P1, P2)
}
\arguments{
\item{P1}{A parray of weights or probabilities}

\item{P2}{A parray of weights or probabilities}
}
\value{
A parray.
}
\description{
Compute the product of two arrays of conditional probabilities or
weights.
}
\details{
Given P1 = P(X|Y,Z) and P2 = P(Y|Z), this function computes the
product P(X,Y|Z) = P(X|Y,Z) P(Y|Z).
}
