% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{ptabs}
\alias{ptabs}
\title{Create a named array of probabilities}
\usage{
ptabs(formula, data, fill = 0)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data frame or named array}

\item{fill}{A vector or named array of replacement values for
division by zero.}
}
\value{
A named array of probabilities
}
\description{
Create a named array of probabilities from a data frame or named
array using a formula interface.
}
\details{
Given a formula of the form \code{w ~ d_1+d_2+...+d_n | c_1+c_2+...+c_m},
\code{ptabs} creates a named array of probabilities of the dependent
variables \code{d_i} conditioned on the conditioning variables \code{c_i}.
If data is a dataframe, the optional response \code{w} specifies a
weighting variable, otherwise it is ignored.  If data is a
parray object, the result is conditioned on both the \code{c_i} and
any existing conditioning variables in \code{data}.

Where conditioning results in division by zero, the undefined
values are replaced with the elements of \code{fill}, which may be a
vector or another named array.
}
