% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parray.R
\name{subarray}
\alias{subarray}
\title{Named indexing}
\usage{
subarray(narray, ...)
}
\arguments{
\item{narray}{A named array}

\item{...}{Variable values to extract}
}
\value{
A subarray
}
\description{
Select a subarray from a named array, indexing by variable names.
}
